backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.6.1
chromadb==0.4.22
sentence-transformers>=2.7.0
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.3
tiktoken==0.5.2
numpy<2

